+++ BattleEngine/src/battleEngine/data/models/Armor.java
@@ -2,7 +2,7 @@ package battleEngine.data.models;
 
 public class Armor {
     public String name;
-    public double armorHP;
+    public int armorHP;
 
     public Armor(String name, int armorHP) {
         this.name = name;
